8 bit data type ID for services

+++ src/canard.h
@@ -194,7 +194,7 @@ int canardBroadcast(CanardInstance* ins, uint64_t data_type_signature, uint16_t 
                     uint8_t priority, const void* payload,
                     uint16_t payload_len);
 int canardRequestOrRespond(CanardInstance* ins, uint8_t destination_node_id, uint64_t data_type_signature,
-                           uint16_t data_type_id, uint8_t* inout_transfer_id,
+                           uint8_t data_type_id, uint8_t* inout_transfer_id,
                            uint8_t priority, CanardRequestResponse kind, const void* payload,
                            uint16_t payload_len);
 const CanardCANFrame* canardPeekTxQueue(const CanardInstance* ins);
